Added udBitField get/set accessors so that position and width of bitfields can be entered into an enumerated type and modified easily Updated udPointCloud to use the udBitField accessors for node index Fixed a couple of warnings
git-svn-id: https://bne-fs-pf-001.euclideon.local/svn/Dev/trunk/src@389 f54097bf-93f5-7042-b87d-fd266665e7ff

+++ udPlatform/Source/udPlatformUtil.cpp
@@ -60,7 +60,7 @@ float udPerfCounterMilliseconds(uint64_t startValue, uint64_t end)
 
   // TODO: Come back and tidy this up to be integer
 
-  double delta = (p.QuadPart - startValue);
+  double delta = (double)(p.QuadPart - startValue);
   
   double ms = (delta) ? (1000.0 / (f.QuadPart / delta)) : 0.0;
   return (float)ms;
